+++ src/condor_schedd.V6/schedd_stats.cpp
@@ -287,16 +287,16 @@ void ScheddStatistics::Publish(ClassAd & ad, const char * config) const
 void ScheddStatistics::Publish(ClassAd & ad, int flags) const
 {
    if ((flags & IF_PUBLEVEL) > 0) {
-      ad.Assign("StatsLifetime", (int)StatsLifetime);
+      ad.Assign("StatsLifetime", StatsLifetime);
       ad.Assign("JobsSizesHistogramBuckets", default_sizes_set);
       ad.Assign("JobsRuntimesHistogramBuckets", default_lifes_set);
       if (flags & IF_VERBOSEPUB)
-         ad.Assign("StatsLastUpdateTime", (int)StatsLastUpdateTime);
+         ad.Assign("StatsLastUpdateTime", StatsLastUpdateTime);
       if (flags & IF_RECENTPUB) {
-         ad.Assign("RecentStatsLifetime", (int)RecentStatsLifetime);
+         ad.Assign("RecentStatsLifetime", RecentStatsLifetime);
          if (flags & IF_VERBOSEPUB) {
-            ad.Assign("RecentWindowMax", (int)RecentWindowMax);
-            ad.Assign("RecentStatsTickTime", (int)RecentStatsTickTime);
+            ad.Assign("RecentWindowMax", RecentWindowMax);
+            ad.Assign("RecentStatsTickTime", RecentStatsTickTime);
          }
       }
    }
